introduces ISimpleIndenter and extracted indenter base class; closes #5911
