+++ Rubberduck.SmartIndenter/AbsoluteCodeLine.cs
@@ -7,7 +7,7 @@ using System.Text.RegularExpressions;
 
 namespace Rubberduck.SmartIndenter
 {
-    internal class AbsoluteCodeLine
+    public class AbsoluteCodeLine
     {
         private const string StupidLineEnding = ": _";
         private static readonly Regex LineNumberRegex = new Regex(@"^(?<number>(-?\d+)|(&H[0-9A-F]{1,8}))(?<separator>:)?\s+(?<code>.*)", RegexOptions.ExplicitCapture | RegexOptions.IgnoreCase);
